# docs/api_references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meeting Minutes Agent API",
    "description": "会议纪要 Agent - 自动生成会议纪要的 AI 服务",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "description": "创建任务\n\n流程:\n1. 验证请求参数\n2. 生成任务 ID\n3. 创建任务记录\n4. 推送到消息队列\n5. 立即返回任务 ID\n\nArgs:\n    request: 创建任务请求\n    user_id: 用户 ID (来自认证)\n    tenant_id: 租户 ID (来自认证)\n    db: 数据库会话\n    queue: 队列管理器\n    \nReturns:\n    CreateTaskResponse: 任务创建响应",
        "operationId": "create_task_api_v1_tasks_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "description": "列出用户的任务\n\nArgs:\n    limit: 返回数量限制\n    offset: 偏移量\n    user_id: 用户 ID (来自认证)\n    db: 数据库会话\n    \nReturns:\n    List[TaskDetailResponse]: 任务列表",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDetailResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Detail",
        "description": "获取任务详情\n\nArgs:\n    task_id: 任务 ID\n    user_id: 用户 ID (来自认证)\n    db: 数据库会话\n    \nReturns:\n    TaskDetailResponse: 任务详细信息\n    \nRaises:\n    HTTPException: 404 如果任务不存在\n    HTTPException: 403 如果无权访问",
        "operationId": "get_task_detail_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete Task",
        "description": "删除任务\n\nArgs:\n    task_id: 任务 ID\n    user_id: 用户 ID (来自认证)\n    db: 数据库会话\n    \nRaises:\n    HTTPException: 404 如果任务不存在\n    HTTPException: 403 如果无权访问",
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "properties": {
          "audio_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Audio Files",
            "description": "音频文件 URL 列表"
          },
          "file_order": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Order",
            "description": "文件排序索引"
          },
          "meeting_type": {
            "type": "string",
            "title": "Meeting Type",
            "description": "会议类型"
          },
          "asr_language": {
            "type": "string",
            "title": "Asr Language",
            "description": "ASR 识别语言",
            "default": "zh-CN+en-US"
          },
          "output_language": {
            "type": "string",
            "title": "Output Language",
            "description": "输出语言",
            "default": "zh-CN"
          },
          "prompt_instance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptInstance"
              },
              {
                "type": "null"
              }
            ],
            "description": "提示词实例"
          },
          "skip_speaker_recognition": {
            "type": "boolean",
            "title": "Skip Speaker Recognition",
            "description": "跳过说话人识别",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "audio_files",
          "meeting_type"
        ],
        "title": "CreateTaskRequest",
        "description": "创建任务请求",
        "example": {
          "asr_language": "zh-CN+en-US",
          "audio_files": [
            "https://tos.example.com/meeting.wav"
          ],
          "file_order": [
            0
          ],
          "meeting_type": "weekly_sync",
          "output_language": "zh-CN",
          "prompt_instance": {
            "language": "zh-CN",
            "parameters": {
              "meeting_description": "产品规划会议"
            },
            "template_id": "tpl_001"
          },
          "skip_speaker_recognition": false
        }
      },
      "CreateTaskResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "任务已创建"
          }
        },
        "type": "object",
        "required": [
          "success",
          "task_id"
        ],
        "title": "CreateTaskResponse",
        "description": "创建任务响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromptInstance": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板 ID"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "语言",
            "default": "zh-CN"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "参数值"
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "PromptInstance",
        "description": "提示词实例",
        "example": {
          "language": "zh-CN",
          "parameters": {
            "meeting_description": "会议标题: 产品规划会议\n会议主题: Q2 产品路线图讨论\n关注重点: 重点关注用户反馈和竞品分析"
          },
          "template_id": "tpl_001"
        }
      },
...
      "TaskDetailResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "meeting_type": {
            "type": "string",
            "title": "Meeting Type"
          },
          "audio_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audio Files"
          },
          "file_order": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "File Order"
          },
          "asr_language": {
            "type": "string",
            "title": "Asr Language"
          },
          "output_language": {
            "type": "string",
            "title": "Output Language"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "error_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "user_id",
          "tenant_id",
          "meeting_type",
          "audio_files",
          "file_order",
          "asr_language",
          "output_language",
          "state",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "TaskDetailResponse",
        "description": "任务详情响应"
      },
      "TaskState": {
        "type": "string",
        "enum": [
          "pending",
          "queued",
          "running",
          "transcribing",
          "identifying",
          "correcting",
          "summarizing",
          "success",
          "failed",
          "partial_success"
        ],
        "title": "TaskState",
        "description": "任务状态枚举"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}